style: lint fix part 2&& optimize types style
