+++ README.md
@@ -15,6 +15,9 @@ completely based on the Web APIs with proper timing using `process.hrtime` or
 
 _In case you need more tiny libraries like tinypool or tinyspy, please consider submitting an [RFC](https://github.com/tinylibs/rfcs)_
 
+## Requirements
+- node version >= v14
+
 ## Installing
 
 ```bash
