+++ src/index.ts
@@ -2,8 +2,6 @@ import Bench from "./bench";
 import Task from "./task";
 
 export { now } from "./utils";
-export type { BenchEvent } from "./event";
-export type { Fn, BenchEvents, TaskEvents, Options, TaskResult, Hook } from "./types";
 
 export { Bench, Task };
 export default Bench;
